Fix clippy warnings and test adjustments
Apply small refactors and test updates to address clippy warnings and clean up test code. Changes include: add #[allow(clippy::items_after_test_module)] to the test module, use to_string() instead of format!(...) for static strings, initialize structs with struct update syntax (..Default::default()) instead of mutating a Default value, pass arrays to std::fs::write instead of slice references, fix an out-of-bounds index calculation, adjust expected numeric literals in tests (e.g. 3.25 and use _usize suffix), and other minor formatting tweaks. All changes are limited to tests and small helper code for lint/correctness improvements.

+++ src/ui/panel_layout.rs
@@ -175,7 +175,7 @@ mod tests {
         pl.set_right_width(999.0);
         assert_eq!(pl.right_physical_width(1.0), MAX_WIDTH as usize);
         pl.set_right_width(350.0);
-        assert_eq!(pl.right_physical_width(1.0), 350 as usize);
+        assert_eq!(pl.right_physical_width(1.0), 350_usize);
     }
 
     #[test]
